temp stopping point, wher is this stat structure?
