+++ CMakeLists.txt
@@ -19,6 +19,7 @@ set(EXTENSION_SOURCES
     src/common/url_utils.cpp
     src/common/schema.cpp
     src/common/iceberg.cpp
+    src/common/request_utils.cpp
     src/iceberg_functions/iceberg_multi_file_reader.cpp
     src/iceberg_functions/iceberg_snapshots.cpp
     src/iceberg_functions/iceberg_scan.cpp
