+++ src/catalog_api.cpp
@@ -3,6 +3,7 @@
 #include "storage/irc_catalog.hpp"
 #include "yyjson.hpp"
 #include "iceberg_utils.hpp"
+#include "request_utils.hpp"
 #include <curl/curl.h>
 #include <sys/stat.h>
 #include <aws/core/Aws.h>
@@ -19,278 +20,7 @@
 using namespace duckdb_yyjson;
 namespace duckdb {
 
-//! We use a global here to store the path that is selected on the ICAPI::InitializeCurl call
-static string SELECTED_CURL_CERT_PATH = "";
 
-static size_t RequestWriteCallback(void *contents, size_t size, size_t nmemb, void *userp) {
-	((std::string *)userp)->append((char *)contents, size * nmemb);
-	return size * nmemb;
-}
-
-// we statically compile in libcurl, which means the cert file location of the build machine is the
-// place curl will look. But not every distro has this file in the same location, so we search a
-// number of common locations and use the first one we find.
-static string certFileLocations[] = {
-        // Arch, Debian-based, Gentoo
-        "/etc/ssl/certs/ca-certificates.crt",
-        // RedHat 7 based
-        "/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem",
-        // Redhat 6 based
-        "/etc/pki/tls/certs/ca-bundle.crt",
-        // OpenSUSE
-        "/etc/ssl/ca-bundle.pem",
-        // Alpine
-        "/etc/ssl/cert.pem"
-};
-
-const string IRCAPI::API_VERSION_1 = "v1";
-
-
-// Look through the the above locations and if one of the files exists, set that as the location curl should use.
-static bool SelectCurlCertPath() {
-	for (string& caFile : certFileLocations) {
-		struct stat buf;
-		if (stat(caFile.c_str(), &buf) == 0) {
-			SELECTED_CURL_CERT_PATH = caFile;
-		}
-	}
-	return false;
-}
-
-static bool SetCurlCAFileInfo(CURL* curl) {
-	if (!SELECTED_CURL_CERT_PATH.empty()) {
-		curl_easy_setopt(curl, CURLOPT_CAINFO, SELECTED_CURL_CERT_PATH.c_str());
-        return true;
-	}
-    return false;
-}
-
-// Note: every curl object we use should set this, because without it some linux distro's may not find the CA certificate.
-static void InitializeCurlObject(CURL * curl, const string &token) {
-  	if (!token.empty()) {
-		curl_easy_setopt(curl, CURLOPT_XOAUTH2_BEARER, token.c_str());
-		curl_easy_setopt(curl, CURLOPT_HTTPAUTH, CURLAUTH_BEARER);
-	}
-    SetCurlCAFileInfo(curl);
-}
-
-static string DeleteRequest(const string &url, const string &token = "", curl_slist *extra_headers = NULL) {
-    CURL *curl;
-    CURLcode res;
-    string readBuffer;
-
-    curl = curl_easy_init();
-    if (curl) {
-        curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, "DELETE");
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, RequestWriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &readBuffer);
-        
-        if(extra_headers) {
-            curl_easy_setopt(curl, CURLOPT_HTTPHEADER, extra_headers);
-        }
-        
-        InitializeCurlObject(curl, token);
-        res = curl_easy_perform(curl);
-        curl_easy_cleanup(curl);
-
-        if (res != CURLcode::CURLE_OK) {
-            string error = curl_easy_strerror(res);
-            throw IOException("Curl DELETE Request to '%s' failed with error: '%s'", url, error);
-        }
-        
-        return readBuffer;
-    }
-    throw InternalException("Failed to initialize curl");
-}
-
-class DuckDBSecretCredentialProvider : public Aws::Auth::AWSCredentialsProviderChain
-{
-public:
-	DuckDBSecretCredentialProvider(const string& key_id, const string &secret, const string &sesh_token) {
-		credentials.SetAWSAccessKeyId(key_id);
-		credentials.SetAWSSecretKey(secret);
-		credentials.SetSessionToken(sesh_token);
-	}
-
-	~DuckDBSecretCredentialProvider() = default;
-
-	Aws::Auth::AWSCredentials GetAWSCredentials() override {
-		return credentials;
-	};
-
-protected:
-	Aws::Auth::AWSCredentials credentials;
-};
-
-static string GetAwsService(const string host) {
-	return host.substr(0, host.find_first_of('.'));
-}
-
-static string GetAwsRegion(const string host) {
-	idx_t first_dot = host.find_first_of('.');
-	idx_t second_dot = host.find_first_of('.', first_dot + 1);
-	return host.substr(first_dot + 1, second_dot - first_dot - 1);
-}
-
-static string GetRequestAws(ClientContext &context, IRCEndpointBuilder endpoint_builder, const string &secret_name) {
-	auto clientConfig = make_uniq<Aws::Client::ClientConfiguration>();
-
-	if (!SELECTED_CURL_CERT_PATH.empty()) {
-		clientConfig->caFile = SELECTED_CURL_CERT_PATH;
-	}
-
-	std::shared_ptr<Aws::Http::HttpClientFactory> MyClientFactory;
-	std::shared_ptr<Aws::Http::HttpClient> MyHttpClient;
-
-	MyHttpClient = Aws::Http::CreateHttpClient(*clientConfig);
-	Aws::Http::URI uri;
-
-	// TODO move this to IRCatalog::GetBaseURL()
-	auto service = GetAwsService(endpoint_builder.GetHost());
-	auto region = GetAwsRegion(endpoint_builder.GetHost());
-
-	// Add iceberg. This is necessary here and should not be included in the host
-	uri.AddPathSegment("iceberg");
-	// push bach the version
-	uri.AddPathSegment(endpoint_builder.GetVersion());
-	// then the warehouse
-	if (service == "glue") {
-		uri.AddPathSegment("catalogs");
-		uri.AddPathSegment(endpoint_builder.GetWarehouse());
-	} else {
-		uri.AddPathSegment(endpoint_builder.GetWarehouse());
-	}
-
-	for (auto &component : endpoint_builder.path_components) {
-		uri.AddPathSegment(component);
-	}
-
-	Aws::Http::Scheme scheme = Aws::Http::Scheme::HTTPS;
-	uri.SetScheme(scheme);
-	// set host
-	uri.SetAuthority(endpoint_builder.GetHost());
-	auto encoded = uri.GetURLEncodedPath();
-
-	const Aws::Http::URI uri_const = Aws::Http::URI(uri);
-	auto create_http_req = Aws::Http::CreateHttpRequest(uri_const,
-									 Aws::Http::HttpMethod::HTTP_GET,
-									 Aws::Utils::Stream::DefaultResponseStreamFactoryMethod);
-
-	std::shared_ptr<Aws::Http::HttpRequest> req(create_http_req);
-
-	// will error if no secret can be found for AWS services
-	auto secret_entry = IRCatalog::GetSecret(context, secret_name);
-	auto kv_secret = dynamic_cast<const KeyValueSecret &>(*secret_entry->secret);
-
-	std::shared_ptr<Aws::Auth::AWSCredentialsProviderChain> provider;
-	provider = std::make_shared<DuckDBSecretCredentialProvider>(
-		kv_secret.secret_map["key_id"].GetValue<string>(),
-		kv_secret.secret_map["secret"].GetValue<string>(),
-		kv_secret.secret_map["session_token"].IsNull() ? "" : kv_secret.secret_map["session_token"].GetValue<string>()
-	);
-	auto signer = make_uniq<Aws::Client::AWSAuthV4Signer>(provider, service.c_str(), region.c_str());
-
-	signer->SignRequest(*req);
-	std::shared_ptr<Aws::Http::HttpResponse> res = MyHttpClient->MakeRequest(req);
-	Aws::Http::HttpResponseCode resCode = res->GetResponseCode();
-	DUCKDB_LOG_DEBUG(context, "iceberg.Catalog.Aws.HTTPRequest", "GET %s (response %d) (signed with key_id '%s' for service '%s', in region '%s')", uri.GetURIString(), resCode, kv_secret.secret_map["key_id"].GetValue<string>(), service.c_str(), region.c_str());
-	if (resCode == Aws::Http::HttpResponseCode::OK) {
-		Aws::StringStream resBody;
-		resBody << res->GetResponseBody().rdbuf();
-		return resBody.str();
-	} else {
-		Aws::StringStream resBody;
-		resBody <<  res->GetResponseBody().rdbuf();
-		throw IOException("Failed to query %s, http error %d thrown. Message: %s", req->GetUri().GetURIString(true), res->GetResponseCode(), resBody.str());
-	}
-}
-
-static string GetRequest(ClientContext &context, const IRCEndpointBuilder &endpoint_builder, const string &secret_name, const string &token = "", curl_slist *extra_headers = NULL) {
-	if (StringUtil::StartsWith(endpoint_builder.GetHost(), "glue." ) || StringUtil::StartsWith(endpoint_builder.GetHost(), "s3tables." )) {
-		auto str = GetRequestAws(context, endpoint_builder, secret_name);
-		return str;
-	}
-	auto url = endpoint_builder.GetURL();
-	CURL *curl;
-	CURLcode res;
-	string readBuffer;
-
-	curl = curl_easy_init();
-	if (curl) {
-		curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-		curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, RequestWriteCallback);
-		curl_easy_setopt(curl, CURLOPT_WRITEDATA, &readBuffer);
-		
-		if(extra_headers) {
-			curl_easy_setopt(curl, CURLOPT_HTTPHEADER, extra_headers);
-		}
-		
-		InitializeCurlObject(curl, token);
-		res = curl_easy_perform(curl);
-		curl_easy_cleanup(curl);
-
-		DUCKDB_LOG_DEBUG(context, "iceberg.Catalog.Curl.HTTPRequest", "GET %s (curl code '%s')", url, curl_easy_strerror(res));
-		if (res != CURLcode::CURLE_OK) {
-			string error = curl_easy_strerror(res);
-			throw IOException("Curl Request to '%s' failed with error: '%s'", url, error);
-		}
-
-		return readBuffer;
-	}
-	throw InternalException("Failed to initialize curl");
-}
-
-static string PostRequest(
-		ClientContext &context,
-		const string &url, 
-		const string &post_data, 
-		const string &content_type = "x-www-form-urlencoded",
-		const string &token = "", 
-		curl_slist *extra_headers = NULL) {
-    string readBuffer;
-    CURL *curl = curl_easy_init();
-    if (!curl) {
-		throw InternalException("Failed to initialize curl");
-	}
-
-	curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-	curl_easy_setopt(curl, CURLOPT_POST, 1L);
-	curl_easy_setopt(curl, CURLOPT_POSTFIELDS, post_data.c_str());
-	curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, RequestWriteCallback);
-	curl_easy_setopt(curl, CURLOPT_WRITEDATA, &readBuffer);
-
-	// Create default headers for content type
-	struct curl_slist *headers = NULL;
-	const string content_type_str = "Content-Type: application/" + content_type;
-	headers = curl_slist_append(headers, content_type_str.c_str());
-	
-	// Append any extra headers
-	if (extra_headers) {
-		struct curl_slist *temp = extra_headers;
-		while (temp) {
-			headers = curl_slist_append(headers, temp->data);
-			temp = temp->next;
-		}
-	}
-	
-	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
-	InitializeCurlObject(curl, token);
-	
-	// Perform the request
-	CURLcode res = curl_easy_perform(curl);
-	
-	// Clean up
-	curl_slist_free_all(headers);
-	curl_easy_cleanup(curl);
-
-	DUCKDB_LOG_DEBUG(context, "iceberg.Catalog.Curl.HTTPRequest", "POST %s (curl code '%s')", url, curl_easy_strerror(res));
-	if (res != CURLcode::CURLE_OK) {
-		string error = curl_easy_strerror(res);
-		throw IOException("Curl Request to '%s' failed with error: '%s'", url, error);
-	}
-	return readBuffer;
-}
 
 static string GetTableMetadata(ClientContext &context, IRCatalog &catalog, const string &schema, const string &table, const string &secret_name) {
 	struct curl_slist *extra_headers = NULL;
@@ -300,7 +30,7 @@ static string GetTableMetadata(ClientContext &context, IRCatalog &catalog, const
 	url.AddPathComponent("tables");
 	url.AddPathComponent(table);
 	extra_headers = curl_slist_append(extra_headers, "X-Iceberg-Access-Delegation: vended-credentials");
-	string api_result = GetRequest(
+	string api_result = RequestUtils::GetRequest(
 		context,
 		url,
 		secret_name,
@@ -325,9 +55,8 @@ static string GetTableMetadataCached(ClientContext &context, IRCatalog &catalog,
 	return GetTableMetadata(context, catalog, schema, table, secret_name);
 }
 
-
 void IRCAPI::InitializeCurl() {
-	SelectCurlCertPath();
+	RequestUtils::SelectCurlCertPath();
 }
 
 vector<string> IRCAPI::GetCatalogs(ClientContext &context, IRCatalog &catalog, IRCCredentials credentials) {
@@ -367,7 +96,7 @@ IRCAPITableCredentials IRCAPI::GetTableCredentials(ClientContext &context, IRCat
 
 string IRCAPI::GetToken(ClientContext &context, string id, string secret, string endpoint) {
 	string post_data = "grant_type=client_credentials&client_id=" + id + "&client_secret=" + secret + "&scope=PRINCIPAL_ROLE:ALL";
-	string api_result = PostRequest(context, endpoint + "/v1/oauth/tokens", post_data);
+	string api_result = RequestUtils::PostRequest(context, endpoint + "/v1/oauth/tokens", post_data);
 	std::unique_ptr<yyjson_doc, YyjsonDocDeleter> doc(ICUtils::api_result_to_doc(api_result));
 	auto *root = yyjson_doc_get_root(doc.get());
 	return IcebergUtils::TryGetStrFromObject(root, "access_token");
@@ -442,7 +171,7 @@ vector<IRCAPITable> IRCAPI::GetTables(ClientContext &context, IRCatalog &catalog
 	url.AddPathComponent("namespaces");
 	url.AddPathComponent(schema);
 	url.AddPathComponent("tables");
-	string api_result = GetRequest(context, url, catalog.secret_name, catalog.credentials.token);
+	string api_result = RequestUtils::GetRequest(context, url, catalog.secret_name, catalog.credentials.token);
 	std::unique_ptr<yyjson_doc, YyjsonDocDeleter> doc(ICUtils::api_result_to_doc(api_result));
 	auto *root = yyjson_doc_get_root(doc.get());
 	auto *tables = yyjson_obj_get(root, "identifiers");
@@ -461,7 +190,7 @@ vector<IRCAPISchema> IRCAPI::GetSchemas(ClientContext &context, IRCatalog &catal
 	auto endpoint_builder = catalog.GetBaseUrl();
 	endpoint_builder.AddPathComponent("namespaces");
 	string api_result =
-	    GetRequest(context, endpoint_builder, catalog.secret_name, catalog.credentials.token);
+	    RequestUtils::GetRequest(context, endpoint_builder, catalog.secret_name, catalog.credentials.token);
 	std::unique_ptr<yyjson_doc, YyjsonDocDeleter> doc(ICUtils::api_result_to_doc(api_result));
 	auto *root = yyjson_doc_get_root(doc.get());
 	auto *schemas = yyjson_obj_get(root, "namespaces");
