+++ src/include/credentials/credential_provider.hpp
@@ -1,0 +1,31 @@
+#include <aws/core/Aws.h>
+#include <aws/s3/S3Client.h>
+#include <aws/core/Aws.h>
+#include <aws/core/auth/AWSCredentials.h>
+#include <aws/core/auth/AWSCredentialsProviderChain.h>
+#include <aws/core/auth/AWSCredentialsProvider.h>
+#include <aws/core/http/HttpClient.h>
+#include <aws/core/http/HttpRequest.h>
+
+namespace duckdb {
+
+class DuckDBSecretCredentialProvider : public Aws::Auth::AWSCredentialsProviderChain
+{
+public:
+	DuckDBSecretCredentialProvider(const string& key_id, const string &secret, const string &sesh_token) {
+		credentials.SetAWSAccessKeyId(key_id);
+		credentials.SetAWSSecretKey(secret);
+		credentials.SetSessionToken(sesh_token);
+	}
+
+	~DuckDBSecretCredentialProvider() = default;
+
+	Aws::Auth::AWSCredentials GetAWSCredentials() override {
+		return credentials;
+	};
+
+protected:
+	Aws::Auth::AWSCredentials credentials;
+};
+
+}
