+++ test/sql/local/irc/test_polaris.test
@@ -1,3 +1,7 @@
+# name: test/sql/local/irc/test_polaris.test
+# description: test integration with iceberg catalog read
+# group: [iceberg]
+
 require parquet
 
 require httpfs
@@ -11,18 +15,22 @@ mode skip
 statement ok
 create secret polaris_secret (
   type s3,
-  KEY_ID 'c1afe0044f87ea17',
-  SECRET 'a65907eb154abc7184d66cca48da91f3'
+  KEY_ID '',
+  SECRET ''
 );
 
-
 statement ok
-attach 'quickstart_catalog' as my_datalake (
+attach 'polaris_demo' as my_datalake (
     type ICEBERG,
     ENDPOINT 'http://0.0.0.0:8181/api/catalog'
-)
-
+);
 
 statement ok
 show all tables;
+
+query II
+select * from namespace.catalog.table;
+----
+1
+2
